Separate classes into their own file + UX struct reorganizing
