+++ Client/Frontend/Home/Library/LibraryShortcutView.swift
