+++ Client/Frontend/Home/TopSites/GoogleTopSiteHelper.swift
